chat_backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0